list.h, hashtable.c
I have uploaded the list.h file as a .h file but I have uploaded the hashtable file as a .c file. I have to modify it a little bit. I will finish the modification by EOD today and replace it with a .h function. And please report any bug in my code it you find any.